requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
aiokafka==0.10.0
SQLAlchemy[asyncio]==2.0.31
asyncpg==0.29.0
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.8.0
openai==1.41.0
tiktoken==0.7.0
